+++ app/models/address.rb
@@ -1,3 +1,4 @@
 class Address < ApplicationRecord
   belongs_to :cart
+  belongs_to :prefecture
 end
