+++ app/models/cart_address.rb
@@ -4,13 +4,12 @@ class CartAddress
 
   with_options presence: true do
     validates :postal_code, format: { with: /\A[0-9]{3}-[0-9]{4}\z/, message: 'is invalid. Include hyphen(-)' }
-    varidates :user_id
-    validates :item_id
     validates :prefecture_id, numericality: { other_than: 1, message: "can't be blank" }
     validates :city
     validates :addresses
     validates :phone_number, numericality: { only_integer: true, message: 'Input only number' }
   end
+
   def save
     cart = Cart.create(item_id: item_id, user_id: user_id)
     Address.create(postal_code: postal_code, prefecture_id: prefecture_id, city: city, addresses: addresses,
