+++ app/models/prefecture.rb
@@ -52,4 +52,5 @@ class Prefecture < ActiveHash::Base
 
   include ActiveHash::Associations
   has_many :items
+  has_many :addreses
 end
